move components into components dir
