+++ code/addons/test/src/components/Panel.test.ts
